+++ modules/src/cep78/constants.rs
@@ -1,3 +1,33 @@
+pub const ACL_PACKAGE_MODE: &str = "acl_package_mode";
+pub const ACL_WHITELIST: &str = "acl_whitelist";
+pub const ALLOW_MINTING: &str = "allow_minting";
+pub const APPROVED: &str = "approved";
+pub const BURN_MODE: &str = "burn_mode";
+pub const BURNT_TOKENS: &str = "burnt_tokens";
+pub const COLLECTION_NAME: &str = "collection_name";
+pub const COLLECTION_SYMBOL: &str = "collection_symbol";
+pub const EVENTS_MODE: &str = "events_mode";
+pub const HOLDER_MODE: &str = "holder_mode";
+pub const IDENTIFIER_MODE: &str = "identifier_mode";
+pub const INSTALLER: &str = "installer";
+pub const JSON_SCHEMA: &str = "json_schema";
+pub const METADATA_MUTABILITY: &str = "metadata_mutability";
+pub const MINTING_MODE: &str = "minting_mode";
+pub const NFT_KIND: &str = "nft_kind";
+pub const NFT_METADATA_KIND: &str = "nft_metadata_kind";
+pub const NFT_METADATA_KINDS: &str = "nft_metadata_kinds";
+pub const NUMBER_OF_MINTED_TOKENS: &str = "number_of_minted_tokens";
+pub const OPERATORS: &str = "operators";
+pub const OPERATOR_BURN_MODE: &str = "operator_burn_mode";
+pub const OWNERSHIP_MODE: &str = "ownership_mode";
+pub const RECEIPT_NAME: &str = "receipt_name";
+pub const REPORTING_MODE: &str = "reporting_mode";
+pub const TOKEN_COUNT: &str = "balances";
+pub const TOKEN_ISSUERS: &str = "token_issuers";
+pub const TOKEN_OWNERS: &str = "token_owners";
+pub const TOTAL_TOKEN_SUPPLY: &str = "total_token_supply";
+pub const TRANSFER_FILTER_CONTRACT: &str = "transfer_filter_contract";
+pub const WHITELIST_MODE: &str = "whitelist_mode";
 pub const PREFIX_PAGE_DICTIONARY: &str = "page";
 
 pub const METADATA_CEP78: &str = "metadata_cep78";
