+++ modules/src/lib.rs
@@ -6,9 +6,9 @@
 extern crate alloc;
 
 pub mod access;
-pub mod cep78;
 pub mod cep18;
 pub mod cep18_token;
+pub mod cep78;
 pub mod erc1155;
 pub mod erc1155_receiver;
 pub mod erc1155_token;
@@ -17,4 +17,5 @@ pub mod erc721;
 pub mod erc721_receiver;
 pub mod erc721_token;
 pub mod security;
+mod storage;
 pub mod wrapped_native;
